Add function to convert ical time strings to UTC
Currently, the function ignores timezone strings that can be supplied with from the ical format. To implement this need to convert lots of strings like 'America/New York' to a UTC offset, or a massive ballache, so pretending everything is UTC for now. Included a test for converting ical time to utc without an offset.